+++ src/routes/changelog/changelog.md
@@ -4,6 +4,7 @@
 - Added difficulties to [classes](/classes)
 - Improve embeds for various pages
 - Recipes now show required adventurer rank
+- Added recipes for consumable items
 
 ## August 15, 2023
 Updated site layout and pages.